Fix: Only add/remove a path at a valid world position.
- This adds the IsVoxelstackInWorld check to the path gui's code,
which was causing the game to crash if a user right-clicked outside
of the world (trying to remove a path at an invalid position).

The GetStack method has assertions for whether or not an x/y position
is within the VoxelWorld. Without checking if the x/y position is
valid before calling this method, the game can crash.